fix(trusty-installer): parse codesign identity from security find-identity output correctly
Closes #2937
Closes #2939
